tracker/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==0.24
Markdown==3.4.4
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
webservices==0.7
zipp==3.16.2
